Fixes #683: Split users into users and presence
