+++ src/users/usersReducers.js
@@ -1,28 +1,14 @@
 /* @flow */
-import { UsersState, Action, ClientPresence, Presence, UserStatus } from '../types';
+import { UsersState, Action } from '../types';
 import {
   LOGOUT,
   LOGIN_SUCCESS,
   ACCOUNT_SWITCH,
-  EVENT_PRESENCE,
   INIT_USERS,
   EVENT_USER_ADD,
   EVENT_USER_REMOVE,
   EVENT_USER_UPDATE,
-  PRESENCE_RESPONSE,
 } from '../actionConstants';
-
-const priorityToState = [
-  'offline',
-  'idle',
-  'active',
-];
-
-const stateToPriority = {
-  offline: 0,
-  idle: 1,
-  active: 2,
-};
 
 const mapApiToStateUser = (user) => ({
   id: user.user_id,
@@ -33,27 +19,6 @@ const mapApiToStateUser = (user) => ({
   isAdmin: user.is_admin,
   isBot: user.is_bot,
 });
-
-export const activityFromPresence = (presence: ClientPresence): UserStatus =>
-  priorityToState[
-    Math.max(...Object.values(presence).map((x: Presence) => stateToPriority[x.status]))
-  ];
-
-export const timestampFromPresence = (presence: ClientPresence): UserStatus =>
-  Math.max(...Object.values(presence).map((x: Presence) => x.timestamp));
-
-export const activityFromTimestamp = (activity: string, timestamp: number) =>
-  ((new Date() / 1000) - timestamp > 60 ? 'offline' : activity);
-
-const updateUserWithPresence = (user: Object, presence: Presence) => {
-  const timestamp = timestampFromPresence(presence);
-
-  return {
-    ...user,
-    status: activityFromTimestamp(activityFromPresence(presence), timestamp),
-    timestamp,
-  };
-};
 
 const initialState: UsersState = [];
 
@@ -67,17 +32,6 @@ export default (state: UsersState = initialState, action: Action): UsersState =>
     case INIT_USERS:
       return action.users.map(mapApiToStateUser);
 
-    case PRESENCE_RESPONSE:
-      return Object.keys(action.presence).reduce((currentState, email) => {
-        const userIndex = state.findIndex(u => u.email === email);
-        if (userIndex === -1) return currentState;
-
-        currentState[userIndex] = // eslint-disable-line
-          updateUserWithPresence(currentState[userIndex], action.presence[email]);
-
-        return currentState;
-      }, [...state]);
-
     case EVENT_USER_ADD:
       return [
         ...state,
@@ -89,17 +43,6 @@ export default (state: UsersState = initialState, action: Action): UsersState =>
 
     case EVENT_USER_UPDATE:
       return state; // TODO
-
-    case EVENT_PRESENCE: {
-      const userIndex = state.findIndex(u => u.email === action.email);
-      if (userIndex === -1) return state;
-
-      const updatedUser = updateUserWithPresence(state[userIndex], action.presence);
-      const newState = [...state];
-      newState[userIndex] = updatedUser;
-
-      return newState;
-    }
 
     default:
       return state;
